Split eenvironment into its own script to continue setup
